Changed initial allPathsFrom() to allPossiblePathsBetween to use as a check to see all the possible paths from one location to another
Utalised Ai to develop a working recursive search algorthim with the in build predicate 'findall(Template, Goal, Bag).'
And incorperated findShortestRoute() with allPathsFrom() [this time with /3 instead of the original /4]
in order to determin the shortest route compared between every path

Also ran a test to see what would happen if no paths were available (by adding locations to the [Start] in the travelPath)
resulting in no path being displayed but distance is displayed as 9999999.
Is possible to use that later in I/O to read out that there was no path found

Also by adding locations to [Start] representing visited locations, this can possibly be used to avoid specific locations if the user wanted to do so:
possible implementation: [Start,Ignore] a new variable that will specified by the user for locations they want to avoid
Path shown would be either impossible or the shortest route in terms of distance, avoiding the locations specified.